test: 값 비교 테스트 추가

+++ backend/src/main/java/com/carffeine/carffeine/station/controller/review/dto/ReviewResponse.java
@@ -6,7 +6,7 @@ public record ReviewResponse(
         Long reviewId,
         Long memberId,
         LocalDateTime latestUpdateDate,
-        long ratings,
+        int ratings,
         String content,
         boolean isUpdated,
         boolean isDeleted
